requirements.txt
chromadb
langchain
openai
google-generativeai
python-dotenv
sentence-transformers
tiktoken
python list_gemini_models.py

python generate_config_gemini.py
